Adding campaigns and mycampaigns view and getting campaigns from the api
Adding the campaigns and my campaign pages and hooked them up to get the campaigns from the backend. Also includes a test page I was working on which would have the campaigns be reordable with a drag and drop but decided to put that on hold for now as it's a bit too complex to be doing right now.